feat: Add feature switch options and an options model for easier factory creation
An options model has been added and the new preferred method for creation is Start(ILogger). This method accepts an optional options model.
One of the new options is IsEnabled, which can be used to toggle off and bypass the performance counter, allowing users to use feature switches to turn performance reporting on and off.